+++ src/tables/mod.rs
@@ -39,6 +39,7 @@ pub mod pclt;
 pub mod post;
 pub mod sbix;
 pub mod stat;
+pub mod svg;
 pub mod vdmx;
 pub mod vhea;
 pub mod vmtx;